scsi: sd: add parameter to stop disks before reboot
I need to use "reboot=p" on my desktop because one of the PCIe devices
does not appear after a warm boot. This results in a very cold boot
because the BIOS turns the PSU off and on.

The scsi sd shutdown process does not send a stop command to disks
before the reboot happens (stop commands are only sent for a shutdown).

The result is that all of my SSDs experience a sudden power loss on
every reboot, which is undesirable behaviour because it could cause data
to be corrupted. These events are recorded in the SMART attributes.

Add a "stop_before_reboot" module parameter that can be used to control
the shutdown behaviour of disks before a reboot. The default will be
the existing behaviour (disks are not stopped).

  sd_mod.stop_before_reboot=<integer>
    0 = disabled (default)
    1 = enabled

The behaviour on shutdown is unchanged: all disks are unconditionally
stopped.

Signed-off-by: Simon Arlott <simon@octiron.net>