x86/reboot/quirks: Add ASRock Z170 Extreme4
If a PCI mode reboot is performed, a power cycle will occur. Automatically
set the reboot quirk for this to prepare for the power off (i.e. stop all
disks).

This will only take effect if PCI mode is manually used. It'll be too late
in the reboot process to prepare for power off if the other reboot methods
fail.

Signed-off-by: Simon Arlott <simon@octiron.net>